Safe extract instead of extractall()
